Use GiB and bytes for sizing tool.

+++ pkg/sizing/algorithm_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func Test_AlgorithTest_Algorith(t *testing.T) {
-	f := func(ingest int) bool {
+	f := func(ingest float64) bool {
 		if ingest < 0 {
 			ingest = -ingest
 		}
